Refactor autorization

+++ Command/QuoteDescription/DeleteByIdCommand.php
@@ -8,7 +8,10 @@ use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Quote\Api\CartRepositoryInterface;
 use Magento\Quote\Model\QuoteRepository;
 use Maisondunet\SaveQuote\Api\Data\QuoteDescriptionInterface;
+use Maisondunet\SaveQuote\Api\Data\QuoteDescriptionInterfaceFactory;
 use Maisondunet\SaveQuote\Api\DeleteQuoteDescriptionByIdInterface;
+use Maisondunet\SaveQuote\Api\QuoteDescriptionAuthorizationInterface;
+use Maisondunet\SaveQuote\Model\NotAuthorizedException;
 use Maisondunet\SaveQuote\Model\QuoteDescriptionModel;
 use Maisondunet\SaveQuote\Model\QuoteDescriptionModelFactory;
 use Maisondunet\SaveQuote\Model\ResourceModel\QuoteDescriptionResource;
@@ -35,6 +38,8 @@ class DeleteByIdCommand implements DeleteQuoteDescriptionByIdInterface
     private QuoteDescriptionResource $resource;
     private QuoteRepository $quoteRepository;
     private CartRepositoryInterface $cartRepository;
+    private QuoteDescriptionAuthorizationInterface $quoteDescriptionAuthorization;
+    private QuoteDescriptionInterfaceFactory $quoteDescriptionInterfaceFactory;
 
     /**
      * @param LoggerInterface $logger
@@ -45,27 +50,31 @@ class DeleteByIdCommand implements DeleteQuoteDescriptionByIdInterface
      */
     public function __construct(
         LoggerInterface              $logger,
+        QuoteDescriptionInterfaceFactory $quoteDescriptionInterfaceFactory,
         QuoteDescriptionModelFactory $modelFactory,
         QuoteDescriptionResource     $resource,
         QuoteRepository $quoteRepository,
-        CartRepositoryInterface $cartRepository
+        CartRepositoryInterface $cartRepository,
+        QuoteDescriptionAuthorizationInterface $quoteDescriptionAuthorization
     ) {
         $this->logger = $logger;
         $this->modelFactory = $modelFactory;
         $this->resource = $resource;
         $this->quoteRepository = $quoteRepository;
         $this->cartRepository = $cartRepository;
+        $this->quoteDescriptionAuthorization = $quoteDescriptionAuthorization;
+        $this->quoteDescriptionInterfaceFactory = $quoteDescriptionInterfaceFactory;
     }
 
     /**
      * @inheritDoc
+     * @throws NotAuthorizedException
      */
     public function execute(int $entityId): void
     {
         try {
-            /** @var QuoteDescriptionModel $model */
             $model = $this->modelFactory->create();
-            $this->resource->load($model, $entityId, QuoteDescriptionInterface::QUOTE_ID);
+            $this->resource->load($model, $entityId);
 
             if (!$model->getData(QuoteDescriptionInterface::QUOTE_ID)) {
                 throw new NoSuchEntityException(
@@ -78,7 +87,14 @@ class DeleteByIdCommand implements DeleteQuoteDescriptionByIdInterface
                 );
             }
 
-            $this->resource->delete($model);
+            $quoteDescription= $this->quoteDescriptionInterfaceFactory->create(["data" => $model->getData()]);
+            if($this->quoteDescriptionAuthorization->canDelete($quoteDescription)) {
+                $this->resource->delete($model);
+            } else {
+                throw new NotAuthorizedException(__("Not allowed"));
+            }
+        } catch (NotAuthorizedException $e) {
+            throw $e;
         } catch (Exception $exception) {
             $this->logger->error(
                 __('Could not delete QuoteDescription. Original message: {message}'),
